==> notebooks/00_prepare_earthquake_groundTruth_USGS.ipynb <==
from usgs_querier import EQFromUSGS
import pandas as pd

def filter_fn(evid):
    if (evid[:2] == 'ci') | (evid[:2]=='nc'):
        return True
    else:
        return False

earthquake_list = []
for t0, t1 in [('2016-02-01', '2017-01-01'),
               ('2017-01-01', '2018-01-01'),
               ('2018-01-01', '2019-01-01'),
               ('2019-01-01', '2020-01-01'),
               ('2020-01-01', '2020-08-01')]:
    earthquake = EQFromUSGS.from_time_range(t0, t1, min_mag=3.5)
    
    earthquake_list.append(earthquake[earthquake['id'].apply(filter_fn)])

df = pd.concat(earthquake_list)
df = df.assign(datetime = pd.to_datetime(df['time']))
df = df.set_index('datetime')

df.to_pickle('USGS_Earthquake_20160201_20200801_M3.0_M3.5.pkl')

## Get smaller events for checking false positive

date_range = pd.date_range('2016-02-01', '2020-08-01', freq='MS')

earthquake_list = []
for i in range(len(date_range)-1):
    t0 = date_range[i]
    t1 = date_range[i+1]
    earthquake = EQFromUSGS.from_time_range(t0, t1, min_mag=2.5, max_mag=3.5)
    earthquake_list.append(earthquake[earthquake['id'].apply(filter_fn)])

df = pd.concat(earthquake_list)
df = df.assign(datetime = pd.to_datetime(df['time']))
df = df.set_index('datetime')

df.to_pickle('USGS_Earthquake_20160201_20200801_M2.5_M3.5.pkl')
