==> notebooks/02_check_events_myshake.ipynb <==
import gzip
import matplotlib.pyplot as plt
import glob
import numpy as np
import datetime
import pandas as pd

plt.style.use('seaborn-poster')
import os

import matplotlib.pyplot as plt

import plotting_utils
import importlib
importlib.reload(plotting_utils)


data_path = '../data/myshake_data/20160212_20170101/*_heartbeat.pkl'

## Read in the EPIC triggers

df_event = pd.read_hdf('../data/event_20160201_20190929.h5')
df_epic = pd.read_hdf('../data/e2trigs_20160201_20170101.h5')

DEBUG = False
output_folder = '../figures/myshake_figures/new_figure_2016/'

plot_simulation = True
for f_hb in glob.glob(data_path):
    
    evid = f_hb.split('/')[-1].split('_')[0]
    
    #if os.path.isfile('../figures/' + evid + '.png'):
    #    continue
    
    if DEBUG:
        output_folder = None
        if evid != 'ci39322287':
            continue
        
    # read in the data
    df_heartbeat = pd.read_pickle(f_hb)
    df_trig = pd.read_pickle(f_hb.replace('_heartbeat', '_trig'))
    earthquake = pd.read_pickle(f_hb.replace('_heartbeat', '_ground_truth'))
    
    # plot the observation 
    plotting_utils.plot_comparison_plot(df_trig,
                                        df_heartbeat,
                                        df_epic,
                                        df_event,
                                        earthquake, 
                                        save_folder=output_folder, 
                                        myshake_legacy=True,
                                        plot_simulation=False,
                                        simulation_path=None)
    # plot the simulation 
    plotting_utils.plot_comparison_plot(df_trig,
                                        df_heartbeat,
                                        df_epic,
                                        df_event,
                                        earthquake, 
                                        save_folder=output_folder, 
                                        myshake_legacy=True,
                                        plot_simulation=plot_simulation,
                                        simulation_path='../data/myshake_data/20160212_20170101/')
