==> notebooks/11_training_a_baseline_model.ipynb <==
import collections
import numpy as np
import matplotlib.pyplot as plt
from sklearn.utils import class_weight
import tensorflow as tf
from tensorflow.keras import datasets, layers, models

data = np.load('./training_small.npz')

model = models.Sequential()
model.add(layers.Conv2D(32, (3, 3), activation='relu', input_shape=(30, 30, 1)))
model.add(layers.MaxPooling2D((2, 2)))
model.add(layers.Conv2D(64, (3, 3), activation='relu'))
model.add(layers.MaxPooling2D((2, 2)))
model.add(layers.Conv2D(64, (3, 3), activation='relu'))

model.add(layers.Flatten())
model.add(layers.Dense(64, activation='relu'))
model.add(layers.Dense(1, activation='sigmoid'))

model.summary()

X = data['X']
y = data['y']
meta = data['meta']
n_quake_trigs = data['n_quake_trigs']

# We replace the nan with zeros
X = np.nan_to_num(X)

ix = np.random.choice(range(len(y)), size=len(y), replace=False)

X = X[:, :, :, np.newaxis][ix]
y = y[ix]
meta = meta[ix]
n_quake_trigs = n_quake_trigs[ix]

collections.Counter(y)

class_weights = class_weight.compute_class_weight('balanced',
                                                 np.unique(y[:50000]),
                                                 y[:50000])

class_weight_dict = dict(enumerate(class_weights))

model.compile(optimizer='adam',
              loss='binary_crossentropy',
              metrics=['accuracy'])

history = model.fit(X[:50000], y[:50000], epochs=10, 
                    validation_split=0.2, batch_size=256, class_weight=class_weight_dict)
