==> notebooks/12_test_on_real_data.ipynb <==
import collections
import glob
import os
import numpy as np
import pickle
import matplotlib.pyplot as plt
from tensorflow.keras import models

import 

model = models.load_model(
    '../deep_learning_trainer/trained_model/baseline/outputBest_acc_model-023-0.989288-0.987414.h5')

myshake_detections = pickle.load(open('../data/myshake_detection_results.pkl', 'rb'))

## Get 

simulation_folder = '../data/simulation_data_pt1p/'
usgs_event_folder = '../data/ground_truth/'
use_old_myshake_detection = True

for simulation_file in glob.glob(os.path.join(simulation_folder, '*')):
    evid = simulation_file.split('/')[-1].split('_')[0]
    ground_truth_file = glob.glob(os.path.join(usgs_event_folder, evid + '*'))[0]
    earthquake = pickle.load(open(ground_truth_file, 'rb'))
    
    simulation_dict = pickle.load(open(simulation_file, 'rb'))
    phones_steady = np.array(simulation_dict[0]['phones_steady'])
    df_trig = simulation_dict[0]['df_trig']
    
    # There are MyShake detections
    if myshake_detections[evid][0][0][0][0]:
        myshake_alert_time_since_origin_sec = myshake_detections[evid][0][1][3]
        break
    else:
        continue
